Azure integration and create event fix

+++ PioneerLigan/Models/EventResult.cs
@@ -6,9 +6,9 @@
         public int EventId { get; set; }
         public int PlayerId { get; set; }
         public int Points { get; set; }
-        public float OMW { get; set; }
-        public float GW { get; set; }
-        public float OGW { get; set; }
+        public decimal OMW { get; set; }
+        public decimal GW { get; set; }
+        public decimal OGW { get; set; }
         public int Placement { get; set; }
         public string PlayerName { get; set; } = String.Empty;
     }
